+++ supervisor/store/__init__.py
@@ -12,7 +12,7 @@ from supervisor.utils.json import read_json_file
 from ..const import REPOSITORY_CORE, REPOSITORY_LOCAL
 from ..coresys import CoreSys, CoreSysAttributes
 from ..exceptions import JsonFileError, StoreError, StoreGitError
-from ..job.decorator import Job, JobCondition
+from ..jobs.decorator import Job, JobCondition
 from .addon import AddonStore
 from .data import StoreData
 from .repository import Repository
